pyproject.toml
[tool.poetry]
name = "mllwebsite"
version = "0.1.0"
description = ""
authors = ["Danial Alihosseini <danial.alihosseini@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.7"
wagtail = "^2.10.1"
wagtailstreamforms = "^3.11"
django-private-storage = "^2.2.2"
python-dotenv = "^0.14.0"
pandas = "^1.1.2"

[tool.poetry.dev-dependencies]
poetry = "^1.0.10"
pylint = "^2.6.0"
autopep8 = "^1.5.4"
rope = "^0.17.0"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
